pyproject.toml
[project]
name = "aide-dashboard"
version = "0.1.0"
description = "AI Developer Effectiveness dashboard — track your AI coding productivity"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "flask>=3.0",
    "pyyaml>=6.0",
]

[project.scripts]
aide = "aide.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aide"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]
